requirements.txt
pybids
rich
pandas
